requirement.txt
# Core
python-dotenv==1.0.0
pandas==2.1.0
httpx==0.25.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # ✅ Ajouter cette ligne

# Data formats
pyarrow==14.0.0  # Pour parquet
fastparquet==2023.10.0

# Orchestration (plus tard)
# apache-airflow==2.7.3

# DBT (plus tard)
# dbt-core==1.6.0
# dbt-postgres==1.6.0
snowflake-connector-python==3.6.0